Enable Perl plugin
Perl plugin allows to use Perl scripts as collectd plugins. This is
usefull as many communautary collectd plugins, developped specifically
for some hardware models, are developped in Perl, mostly for
maintainibility, flexibility and tunability reasons.

This only adds a recommend to libperl on collectd binary package and a
suggest on libperl for collectd-core binary package. Then, it should
not hurt Scibian HPC clusters.